기능추가: 온,습도 입력 및 유니티 전달 기능 추가
온습도를 다이얼로그로 입력해서 식물 성장 부분에서
사용하고, 유니티로 두 파라미터를 전달하는 기능을
추가했습니다.

+++ WeatherGarden/app/src/main/java/com/example/weathergarden/garden/GroundInfo.java
@@ -24,7 +24,7 @@ public class GroundInfo implements Serializable {
     @ColumnInfo(name = "plant_code") @NonNull
     public String plantCode;
     @ColumnInfo(name = "water") @NonNull
-    public int water;
+    public float water;
     @ColumnInfo(name = "nutrient") @NonNull
     public int nutrient;
     @ColumnInfo(name = "wither") @NonNull
